changed prefix to pfs, added discovery client to config server
