+++ pfs-notification/src/test/java/ru/gd/dev/spring/pi/notification/PfsNotificationApplicationTest.java
@@ -1,0 +1,15 @@
+package ru.gd.dev.spring.pi.notification;
+
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+/**
+ * @author V. Zinchenko
+ */
+public class PfsNotificationApplicationTest {
+
+    @Test
+    public void contextLoads() {
+    }
+}
