+++ pfs-server-config/src/test/java/ru/gd/dev/spring/pfs/server/config/PfsServerConfigApplicationTest.java
@@ -1,0 +1,17 @@
+package ru.gd.dev.spring.pfs.server.config;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class PfsServerConfigApplicationTest {
+
+    @Test
+    public void contextLoads() {
+    }
+
+}
+
