+++ pfs-server-discovery/src/main/resources/application.yml
