+++ pfs-server-discovery/src/test/java/ru/gb/dev/spring/pfs/server/discovery/PfsServerDiscoveryApplicationTest.java
@@ -1,0 +1,17 @@
+package ru.gb.dev.spring.pfs.server.discovery;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class PfsServerDiscoveryApplicationTest {
+
+    @Test
+    public void contextLoads() {
+    }
+
+}
+
