+++ pfs-server-gateway/src/main/java/ru/gb/dev/spring/pfs/server/gateway/PfsServerGatewayApplication.java
@@ -1,0 +1,18 @@
+package ru.gb.dev.spring.pfs.server.gateway;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.cloud.netflix.zuul.EnableZuulProxy;
+
+@EnableZuulProxy
+@EnableDiscoveryClient
+@SpringBootApplication
+public class PfsServerGatewayApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(PfsServerGatewayApplication.class, args);
+    }
+
+}
+
