+++ pfs-server-gateway/src/test/java/ru/gb/dev/spring/pfs/server/gateway/PfsServerGatewayApplicationTest.java
@@ -1,0 +1,17 @@
+package ru.gb.dev.spring.pfs.server.gateway;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class PfsServerGatewayApplicationTest {
+
+    @Test
+    public void contextLoads() {
+    }
+
+}
+
