--- pi-accounting/src/test/java/ru/gb/dev/spring/piaccounting/PiAccountingApplicationTests.java
@@ -1,17 +1,0 @@
-package ru.gb.dev.spring.piaccounting;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class PiAccountingApplicationTests {
-
-    @Test
-    public void contextLoads() {
-    }
-
-}
-
