--- pi-server-config/src/test/java/ru/gd/dev/spring/pi/server/config/PiServerConfigApplicationTests.java
@@ -1,17 +1,0 @@
-package ru.gd.dev.spring.pi.server.config;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class PiServerConfigApplicationTests {
-
-    @Test
-    public void contextLoads() {
-    }
-
-}
-
