--- pi-server-gateway/src/main/java/ru/gb/dev/spring/pi/server/gateway/PiServerGatewayApplication.java
@@ -1,18 +1,0 @@
-package ru.gb.dev.spring.pi.server.gateway;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.cloud.netflix.zuul.EnableZuulProxy;
-
-@EnableZuulProxy
-@EnableDiscoveryClient
-@SpringBootApplication
-public class PiServerGatewayApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(PiServerGatewayApplication.class, args);
-    }
-
-}
-
